Move an OutsideClick component into common components and use it in the Withdraw currency list
