+++ src/front/shared/components/OutsideClick/index.tsx
