+++ src/front/shared/components/ui/DropDown/index.tsx
@@ -6,7 +6,7 @@ import Link from 'local_modules/sw-valuelink'
 import { constants } from 'helpers'
 import { FormattedMessage } from 'react-intl'
 import Input from 'components/forms/Input/Input'
-import OutsideClick from './OutsideClick'
+import OutsideClick from 'components/OutsideClick'
 
 const isDark = localStorage.getItem(constants.localStorage.isDark)
